requirements.txt
streamlit
pandas
numpy
spacy
nltk
PyPDF2
python-docx
scikit-learn
